+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - Unreleased
+### Added
+- AstarteVoyagerIngress now has two more options in `api`: `serveMetrics` and `serveMetricsToSubnet`, to
+  give fine-grained control on who has access to `/metrics`
+
+### Changed
+- All `/metrics` endpoints are no longer exposed by default
+
 ## [0.11.0-rc.1] - 2020-03-26
 ### Fixed
 - Allocate resources correctly in Components when non-explicit per-component requirements are given